Move image metadata to a separate build step
There's a bug where the image dimensions don't take into account the
EXIF orientation. From
https://github.com/python-pillow/Pillow/discussions/9115 it looks like
fixing that will mean reading all the image data, not just headers. For
performance, I think that should be done in a separate build step from
rendering templates.

This commit does not fix that bug, but should make it easier to fix it
in another commit.